Refactored to ManageResultDataService
