+++ Source/Microsoft.Teams.Apps.CompanyCommunicator.Send.Func/Startup.cs
@@ -17,6 +17,7 @@ namespace Microsoft.Teams.Apps.CompanyCommunicator.Send.Func
     using Microsoft.Teams.Apps.CompanyCommunicator.Common.Services.MessageQueue;
     using Microsoft.Teams.Apps.CompanyCommunicator.Send.Func.Services.AccessTokenServices;
     using Microsoft.Teams.Apps.CompanyCommunicator.Send.Func.Services.ConversationServices;
+    using Microsoft.Teams.Apps.CompanyCommunicator.Send.Func.Services.DataServices;
     using Microsoft.Teams.Apps.CompanyCommunicator.Send.Func.Services.NotificationServices;
 
     /// <summary>
@@ -53,7 +54,7 @@ namespace Microsoft.Teams.Apps.CompanyCommunicator.Send.Func
             builder.Services.AddTransient<CreateUserConversationService>();
             builder.Services.AddTransient<SendNotificationService>();
             builder.Services.AddTransient<DelaySendingNotificationService>();
-            builder.Services.AddTransient<ManageNotificationResultService>();
+            builder.Services.AddTransient<ManageResultDataService>();
         }
     }
 }
